requirements.txt
absl-py==0.9.0
abstract==2020.7.26
aiodns==2.0.0
aiohttp==3.7.3
aiohttp-socks==0.4.1
algorithms==0.1.4
anaconda==0.0.1.1
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.4
astroid==2.5.2
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
base32hex==1.0.2
beautifulsoup4==4.9.3
bert==2.2.0
bert-for-tf2==0.14.9
bert-tensorflow==1.0.4
bleach==3.2.1
bz2file==0.98
cachetools==4.1.1
cchardet==2.1.7
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colouration==2020.7.26
common==0.1.2
config==0.5.0.post0
coverage==5.3
cryptography==3.4.7
cssselect==1.1.0
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
Django==2.2
Django-2.2.19==2.2.19
django-cors-headers==3.7.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
elasticsearch==7.10.0
emoji==1.2.0
entrypoints==0.3
erlastic==2.0.0
external==0.0.1
fake-useragent==0.1.11
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
Flask==1.0.2
Flask-Cors==3.0.10
fuzzywuzzy==0.18.0
gast==0.3.3
gensim==3.8.0
geographiclib==1.50
geopy==2.0.0
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googletransx==2.4.2
graphviz==0.15
grpcio==1.34.0
h5py==2.10.0
idna==2.8
iniconfig==1.1.1
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.8.0
itsdangerous==1.1.0
jedi==0.18.0
jieba==0.42.1
jieba3k==0.35.1
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
lxml==4.6.3
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.5
newspaper3k==0.2.8
nltk==3.5
notebook==6.2.0
numpy==1.18.0
oauthlib==3.1.0
OpenCC==1.1.2
opencc-python-reimplemented==0.1.6
opt-einsum==3.3.0
optimization==0.0.1
packaging==20.7
pandas==1.2.3
pandocfilters==1.4.3
params-flow==0.8.2
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.10
protobuf==3.14.0
ptyprocess==0.7.0
py==1.9.0
py-params==0.10.2
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycparser==2.20
Pygments==2.7.4
PyJWT==2.0.1
pylint==2.7.4
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pyspark==3.0.1
pytest==6.1.2
pytest-cov==2.10.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
pyzmq==21.0.0
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.22.0
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa==4.6
schedule==0.6.0
scikit-learn==0.24.1
scipy==1.6.2
self==2020.12.3
Send2Trash==1.5.0
sgmllib3k==1.0.0
six==1.15.0
sklearn==0.0
smart-open==4.0.1
smokesignal==0.7.0
soupsieve==2.0.1
sqlparse==0.4.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
textrank4zh==0.3
Theano==1.0.5
threadpoolctl==2.1.0
tinysegmenter==0.3
tldextract==3.1.0
tokenization==1.0.7
toml==0.10.2
torch==1.8.0
tornado==6.1
tqdm==4.32.2
traitlets==5.0.5
twint @ file:///Users/justforstar/Documents/2J301/JG_Project/twint
typing-extensions==3.7.4.3
urllib3==1.25.11
utils==1.0.1
virtualenv==20.4.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yapf==0.31.0
yarl==1.6.3
